sgra_simple_rocket separation
In order to allow collaborative development of restoration and gradient
functions, these were separated into grad_sgra and rest_sgra. The main
function is, appropriately, main_sgra. prob_rocket_sgra and itsme
continue the same.

In the quest for a good restoration, a numerical instability was found
in the solution for the co-states lambda. This still has to be overcome.
